fixed large file processing

+++ libs/src/csharp/SalmonFS/Utils/FileImporter.cs
@@ -249,7 +249,7 @@ public abstract class FileImporter
 
             byte[] bytes = new byte[bufferSize];
             int bytesRead;
-            while ((bytesRead = sourceStream.Read(bytes, 0, Math.Min(bytes.Length, (int)(count - totalPartBytesRead)))) > 0
+            while ((bytesRead = sourceStream.Read(bytes, 0, (int)Math.Min((long)bytes.Length, count - totalPartBytesRead))) > 0
                     && totalPartBytesRead < count)
             {
                 if (stopped)
